+++ index.js
@@ -7,6 +7,19 @@ module.exports.escape = function (arg) {
   return main.escapeShellArgByPlatform(arg, platform);
 };
 
+module.exports.escapeAll = function (args) {
+  if (!Array.isArray(args)) args = [args];
+
+  const platform = os.platform();
+  const result = [];
+  for (const arg of args) {
+    const safeArg = main.escapeShellArgByPlatform(arg, platform);
+    result.push(safeArg);
+  }
+
+  return result;
+};
+
 module.exports.quote = function (arg) {
   const platform = os.platform();
   return main.quoteByPlatform(arg, platform);